skills/medium-term/skill8-performance-prophet/requirements.txt
cProfile>=0.0.0
line_profiler>=4.1.0
memory_profiler>=0.61.0
psutil>=5.9.6
py-spy>=0.3.14
snakeviz>=2.2.0
gprof2dot>=2022.7.29
sqlalchemy>=2.0.23
sqlparse>=0.4.4
numpy>=1.26.2
pandas>=2.1.3
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.15.0
locust>=2.18.4
pytest-benchmark>=4.0.0
pyyaml>=6.0.1
rich>=13.7.0
click>=8.1.7
pydantic>=2.5.0
scikit-learn>=1.3.2
scipy>=1.11.4